+++ packages/common/src/utils/assertUnreachable.ts
@@ -1,8 +1,12 @@
 const assertUnreachable = (
     _x: never,
-    error: string = "Didn't expect to get here",
+    error: string | Error = "Didn't expect to get here",
 ): never => {
-    throw new Error(error);
+    if (typeof error === 'string') {
+        throw Error(error);
+    } else {
+        throw error;
+    }
 };
 
 export default assertUnreachable;