+++ packages/frontend/src/components/ChartConfigPanel/Series/SingleSeriesConfiguration.tsx
@@ -1,11 +1,14 @@
-import { Button, InputGroup, Switch } from '@blueprintjs/core';
+import { Button, Icon, InputGroup, Switch } from '@blueprintjs/core';
 import {
     CartesianChartLayout,
     CartesianSeriesType,
+    getSeriesId,
     Series,
 } from '@lightdash/common';
 import React, { FC } from 'react';
+import { DraggableProvidedDragHandleProps } from 'react-beautiful-dnd';
 import {
+    DragIcon,
     SeriesExtraInputs,
     SeriesExtraInputWrapper,
     SeriesExtraSelect,
@@ -26,6 +29,7 @@ type Props = {
     updateSingleSeries: (updatedSeries: Series) => void;
     isOpen?: boolean;
     toggleIsOpen?: () => void;
+    dragHandleProps?: DraggableProvidedDragHandleProps;
 };
 
 const SingleSeriesConfiguration: FC<Props> = ({
@@ -39,6 +43,7 @@ const SingleSeriesConfiguration: FC<Props> = ({
     isSingle,
     isOpen,
     toggleIsOpen,
+    dragHandleProps,
 }) => {
     const type =
         series.type === CartesianSeriesType.LINE && !!series.areaStyle
@@ -47,6 +52,13 @@ const SingleSeriesConfiguration: FC<Props> = ({
     return (
         <SeriesWrapper $isSingle={isSingle}>
             <SeriesMainInputs $isGrouped={isGrouped}>
+                {isGrouped && (
+                    <DragIcon
+                        tagName="div"
+                        icon="drag-handle-vertical"
+                        {...dragHandleProps}
+                    />
+                )}
                 <SeriesColorPicker
                     color={series.color || fallbackColor}
                     onChange={(color) => {